+++ ChangeLog.md
@@ -12,6 +12,7 @@ Eventum Issue Tracking System
 - Convert report pages to Controller logic (@glensc, #129)
 - Move bin script to use Command class (@glensc, #137)
 - Move Db classes under Eventum\Db namespace (@glensc, #136)
+- Add $scm parameter to handleSCMCheckins workflow method (@glensc)
 
 2016-01-18, Version [3.0.8]
 ---------------------------